Add Contato CRUD

+++ petshop-backend/src/main/java/com/example/petshop/controllers/ClienteController.java
@@ -51,6 +51,11 @@ public class ClienteController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity deleteCliente(@PathVariable Long id){
+        Optional<ClienteEntity> cliente = clienteService.getClienteById(id);
+        if (cliente.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
         clienteService.deleteCliente(id);
         return ResponseEntity.ok(null);
     }
